svn path=/trunk/mods/zk/; revision=12622

+++ LuaUI/Widgets/chili_new/controls/control.lua
@@ -69,12 +69,14 @@ Control = Object:Inherit{
     pressed  = false,
     enabled  = true, --FIXME implement
   },
-
   skin            = nil,
   skinName        = nil,
-
+  
   drawcontrolv2 = nil, --// disable backward support with old DrawControl gl state (with 2.1 self.xy translation isn't needed anymore)
   noSelfHitTest = nil,
+
+  useRTT = ((gl.CreateFBO and gl.BlendFuncSeparate) ~= nil),
+  useDLists = (gl.CreateList ~= nil) and false, --FIXME broken in combination with RTT (wrong blending)
   
   OnResize        = {},
 }
@@ -886,7 +888,11 @@ end
 --//=============================================================================
 
 function Control:_CheckIfRTTisAppreciated()
-	if self._cantUseRTT then
+	if (self.width <= 0)or(self.height <= 0) then
+		return false
+	end
+
+	if self._cantUseRTT or not(self.useRTT) then
 		return false
 	end
 
@@ -907,12 +913,10 @@ end
 function Control:_UpdateOwnDList()
 	if not self.parent then return end
 	if not self:IsInView() then return end
-
-	self:CallChildren('_UpdateOwnDList')
+	if not self.useDLists then return end
 
 	gl.DeleteList(self._own_dlist)
-	--self._own_dlist = nil
-	--self._own_dlist = gl.CreateList(self.DrawControl, self)
+	self._own_dlist = gl.CreateList(self.DrawControl, self)
 end
 
 
@@ -925,15 +929,21 @@ function Control:_UpdateChildrenDList()
 		if (contentWidth <= 0)or(contentHeight <= 0) then return end
 		self:CreateViewTexture("children", contentWidth, contentHeight, self.DrawChildrenForList, self, true)
 	end
+
+	--FIXME
+	--if self.useDLists then
+	--	self._children_dlist = gl.CreateList(self.DrawChildrenForList, self, true)
+	--end
 end
 
 
 function Control:_UpdateAllDList()
 	if not self.parent then return end
 	if not self:IsInView() then return end
-	if (self.width <= 0)or(self.height <= 0) then return end
 
 	local RTT = self:_CheckIfRTTisAppreciated()
+
+	gl.DeleteList(self._all_dlist)
 
 	if RTT then
 		self._usingRTT = true
@@ -945,19 +955,23 @@ function Control:_UpdateAllDList()
 		local fboName = "_fbo_" .. suffix_name
 		local texw = "_texw_" .. suffix_name
 		local texh = "_texh_" .. suffix_name
-		gl.DeleteFBO(self[fboName])
-		gl.DeleteTexture(self[texname])
-		gl.DeleteRBO(self[texStencilName])
+		if gl.DeleteFBO then
+			gl.DeleteFBO(self[fboName])
+			gl.DeleteTexture(self[texname])
+			gl.DeleteRBO(self[texStencilName])
+		end
 		self[texStencilName] = nil
 		self[texname] = nil
 		self[fboName] = nil
 		self[texw] = nil
 		self[texh] = nil
 		self._usingRTT = false
-	end
 
-	--gl.DeleteList(self._all_dlist)
-	--self._all_dlist = gl.CreateList(self.DrawForList,self)
+		--FIXME
+		--if self.useDLists then
+		--	self._all_dlist = gl.CreateList(self.DrawForList,self)
+		--end
+	end
 
 	if (self.parent)and(not self.parent._needRedraw)and(self.parent._UpdateAllDList) then
 		TaskHandler.RequestInstantUpdate(self.parent)
@@ -1251,7 +1265,6 @@ end
 function Control:Draw()
 	self._redrawCounter = (self._redrawCounter or 0) + 1
 	if (not self._in_update and not self._usingRTT and self:_CheckIfRTTisAppreciated()) then self:InvalidateSelf() end
-	if (self.width <= 0)or(self.height <= 0) then return end
 
 	if (self._tex_all) then
 		gl.PushMatrix()
